Refine response message
