+++ src/SSW.SophieBot/components/SSWSophieBot.HttpClientComponents.PersonQuery/Actions/GetEmployeesByDateAction.cs
@@ -36,32 +36,12 @@ namespace SSWSophieBot.HttpClientComponents.PersonQuery.Actions
 
         public override async Task<DialogTurnResult> BeginDialogAsync(DialogContext dc, object options = null, CancellationToken cancellationToken = default)
         {
-            static GetAppointmentModel GetAppointmentBy(DateTime date, List<GetAppointmentModel> appointments)
-            {
-                var leavePhrases = new string[] { "annual leave", "non working", "non-working", "leave", "holiday", "time in lieu", "hour leave", "hours leave", "day off", "days off" };
-                var results = appointments
-                    .Where(appointment => date.Ticks >= GetTicksFrom(appointment.Start) && date.Ticks <= GetTicksFrom(appointment.End))
-                    .Where(appointment => !leavePhrases.Any(appointment.Subject.ToLower().Contains))
-                    .ToList();
-                return results.Count != 0 ? results[0] : null;
-            }
-
-            static long GetTicksFrom(DateTimeOffset date)
-            {
-                return date.UtcDateTime.Ticks;
-            }
-
-            DateTime ToUserLocalTime(DateTime dateTime)
-            {
-                var serverLocalTime = DateTime.SpecifyKind(dateTime, DateTimeKind.Utc);
-                var utcOffset = dc.Context.Activity.LocalTimestamp.GetValueOrDefault().Offset;
-                return serverLocalTime.Subtract(utcOffset);
-            }
-
             var employees = dc.GetValue(Employees);
             var dateString = dc.GetValue(Date);
 
-            var date = dateString != null && dateString != "" ? ToUserLocalTime(DateTime.Parse(dateString)).AddHours(9) : DateTime.Now.ToUniversalTime();
+            var date = dateString != null && dateString != "" 
+                ? EmployeesHelper.ToUserLocalTime(dc, DateTime.Parse(dateString)).AddHours(9) 
+                : DateTime.Now.ToUniversalTime();
 
             var result = employees.Select(e => new EmployeeByDateModel
             {
@@ -69,11 +49,11 @@ namespace SSWSophieBot.HttpClientComponents.PersonQuery.Actions
                 DefaultSite = e.DefaultSite,
                 AvatarUrl = e.AvatarUrl,
                 DisplayName = $"{e.FirstName} {e.LastName}",
-                CurrentAppointment = GetAppointmentBy(date, e.Appointments),
+                CurrentAppointment = EmployeesHelper.GetAppointmentBy(date, e.Appointments),
                 Title = e.Title,
             })
-                .Where(e => e.CurrentAppointment != null)
-                .ToList();
+            .Where(e => e.CurrentAppointment != null)
+            .ToList();
 
 
             if (Result != null)
