+++ src/SSW.SophieBot/components/SSWSophieBot.HttpClientComponents.PersonQuery/Models/EmployeeBillableItemModel.cs
@@ -13,6 +13,9 @@ namespace SSWSophieBot.HttpClientComponents.PersonQuery.Models
         [JsonProperty("billedDays")]
         public int BilledDays { get; set; }
 
+        [JsonProperty("onClientWork")]
+        public bool OnClientWork { get; set; }
+
         [JsonProperty("lastSeen")]
         public string LastSeen { get; set; }
     }
